[types] merge account_state_blob mod
No real point for having this in a separate directory. Let's make this
rusty...
